refactor: 수신 클래스 kitchen 모듈로 이동
