+++ tacocloud-kitchen/src/main/java/tacos/messaging/jms/JmsOrderReceiver.java
@@ -1,9 +1,10 @@
-package tacos.receive;
+package tacos.messaging.jms;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jms.core.JmsTemplate;
 import org.springframework.stereotype.Component;
 import tacos.Order;
+import tacos.OrderReceiver;
 
 @Component
 public class JmsOrderReceiver implements OrderReceiver {
